Adjust memory chart and relocate Samba management
