+++ src/pages/Users.tsx
@@ -11,7 +11,6 @@ import {
   type ChangeEvent,
   type SyntheticEvent,
   useCallback,
-  useEffect,
   useMemo,
   useState,
 } from 'react';
@@ -21,9 +20,6 @@ import TabPanel from '../components/TabPanel';
 import OsUserCreateModal from '../components/users/OsUserCreateModal';
 import OsUsersTable from '../components/users/OsUsersTable';
 import SambaUserCreateModal from '../components/users/SambaUserCreateModal';
-import SambaUserPasswordModal from '../components/users/SambaUserPasswordModal';
-import SambaUsersTable from '../components/users/SambaUsersTable';
-import SelectedSambaUsersDetailsPanel from '../components/users/SelectedSambaUsersDetailsPanel';
 import {
   DEFAULT_LOGIN_SHELL,
   USERS_TABS,
@@ -31,10 +27,8 @@ import {
 } from '../constants/users';
 import { useCreateOsUser } from '../hooks/useCreateOsUser';
 import { useCreateSambaUser } from '../hooks/useCreateSambaUser';
-import { useEnableSambaUser } from '../hooks/useEnableSambaUser';
 import { useOsUsers } from '../hooks/useOsUsers';
 import { useSambaUsers } from '../hooks/useSambaUsers';
-import { useUpdateSambaUserPassword } from '../hooks/useUpdateSambaUserPassword';
 import { normalizeOsUsers } from '../utils/osUsers';
 import { normalizeSambaUsers } from '../utils/sambaUsers';
 
@@ -48,20 +42,6 @@ const Users = () => {
   const [sambaCreateInitialUsername, setSambaCreateInitialUsername] = useState<
     string | undefined
   >(undefined);
-  const [selectedSambaUsers, setSelectedSambaUsers] = useState<string[]>([]);
-  const [isPasswordModalOpen, setIsPasswordModalOpen] = useState(false);
-  const [passwordModalUsername, setPasswordModalUsername] = useState<
-    string | null
-  >(null);
-  const [passwordModalError, setPasswordModalError] = useState<string | null>(
-    null
-  );
-  const [pendingEnableUsername, setPendingEnableUsername] = useState<
-    string | null
-  >(null);
-  const [pendingPasswordUsername, setPendingPasswordUsername] = useState<
-    string | null
-  >(null);
 
   const osUsersQuery = useOsUsers({
     includeSystem,
@@ -120,28 +100,6 @@ const Users = () => {
     onError: (message) => {
       setSambaCreateError(message);
       toast.error(`ایجاد کاربر Samba با خطا مواجه شد: ${message}`);
-    },
-  });
-
-  const enableSambaUser = useEnableSambaUser({
-    onSuccess: (username) => {
-      toast.success(`کاربر ${username} فعال شد.`);
-    },
-    onError: (message) => {
-      toast.error(`فعال‌سازی کاربر با خطا مواجه شد: ${message}`);
-    },
-  });
-
-  const updateSambaPassword = useUpdateSambaUserPassword({
-    onSuccess: (username) => {
-      toast.success(`گذرواژه کاربر ${username} بروزرسانی شد.`);
-      setIsPasswordModalOpen(false);
-      setPasswordModalUsername(null);
-      setPasswordModalError(null);
-    },
-    onError: (message) => {
-      setPasswordModalError(message);
-      toast.error(`تغییر گذرواژه با خطا مواجه شد: ${message}`);
     },
   });
 
@@ -233,89 +191,6 @@ const Users = () => {
     [createOsUser, createSambaUser]
   );
 
-  const handleToggleSelectSambaUser = useCallback(
-    (user: { username: string }, checked: boolean) => {
-      setSelectedSambaUsers((prev) => {
-        if (checked) {
-          if (prev.includes(user.username)) {
-            return prev;
-          }
-
-          return [...prev, user.username];
-        }
-
-        return prev.filter((username) => username !== user.username);
-      });
-    },
-    []
-  );
-
-  const handleRemoveSelectedSambaUser = useCallback((username: string) => {
-    setSelectedSambaUsers((prev) => prev.filter((item) => item !== username));
-  }, []);
-
-  const handleEnableSambaUser = useCallback(
-    (user: { username: string }) => {
-      setPendingEnableUsername(user.username);
-      enableSambaUser.mutate(
-        { username: user.username },
-        {
-          onSettled: () => {
-            setPendingEnableUsername(null);
-          },
-        }
-      );
-    },
-    [enableSambaUser]
-  );
-
-  const handleOpenPasswordModal = useCallback(
-    (user: { username: string }) => {
-      setPasswordModalUsername(user.username);
-      setPasswordModalError(null);
-      updateSambaPassword.reset();
-      setIsPasswordModalOpen(true);
-    },
-    [updateSambaPassword]
-  );
-
-  const handleClosePasswordModal = useCallback(() => {
-    setIsPasswordModalOpen(false);
-    setPasswordModalUsername(null);
-    setPasswordModalError(null);
-    updateSambaPassword.reset();
-  }, [updateSambaPassword]);
-
-  const handleSubmitPasswordChange = useCallback(
-    (payload: { username: string; new_password: string }) => {
-      setPendingPasswordUsername(payload.username);
-      updateSambaPassword.mutate(payload, {
-        onSettled: () => {
-          setPendingPasswordUsername(null);
-        },
-      });
-    },
-    [updateSambaPassword]
-  );
-
-  useEffect(() => {
-    setSelectedSambaUsers((prev) =>
-      prev.filter((username) =>
-        sambaUsers.some((user) => user.username === username)
-      )
-    );
-  }, [sambaUsers]);
-
-  const selectedSambaUserItems = useMemo(
-    () =>
-      selectedSambaUsers
-        .map((username) =>
-          sambaUsers.find((user) => user.username === username)
-        )
-        .filter((user): user is (typeof sambaUsers)[number] => Boolean(user)),
-    [sambaUsers, selectedSambaUsers]
-  );
-
   return (
     <Box
       sx={{
@@ -350,7 +225,6 @@ const Users = () => {
         }}
       >
         <Tab label="کاربران سامانه" value={USERS_TABS.os} />
-        <Tab label="کاربران samba" value={USERS_TABS.samba} />
         <Tab label="سایر کاربران" value={USERS_TABS.other} />
       </Tabs>
 
@@ -423,74 +297,6 @@ const Users = () => {
         </Box>
       </TabPanel>
 
-      <TabPanel value={USERS_TABS.samba} currentValue={activeTab}>
-        <Box sx={{ display: 'flex', flexDirection: 'column', gap: 3 }}>
-          <Box
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between',
-              gap: 2,
-              flexWrap: 'wrap',
-            }}
-          >
-            <Typography
-              variant="h6"
-              sx={{ color: 'var(--color-primary)', fontWeight: 700 }}
-            >
-              مدیریت کاربران Samba
-            </Typography>
-
-            <Button
-              onClick={() => handleOpenSambaCreateModal()}
-              variant="contained"
-              sx={{
-                px: 3,
-                py: 1.25,
-                borderRadius: '3px',
-                fontWeight: 700,
-                fontSize: '0.95rem',
-                background:
-                  'linear-gradient(135deg, var(--color-primary) 0%, var(--color-primary-light) 100%)',
-                color: 'var(--color-bg)',
-                boxShadow:
-                  '0 16px 32px -18px color-mix(in srgb, var(--color-primary) 55%, transparent)',
-                '&:hover': {
-                  background:
-                    'linear-gradient(135deg, var(--color-primary-light) 0%, var(--color-primary) 100%)',
-                },
-              }}
-            >
-              افزودن کاربر Samba
-            </Button>
-          </Box>
-
-          {/*<Typography variant="body2" sx={{ color: 'var(--color-secondary)' }}>*/}
-          {/*  از جدول زیر می‌توانید کاربران Samba را مشاهده، برای مقایسه انتخاب و*/}
-          {/*  عملیات فعال‌سازی یا تغییر گذرواژه را انجام دهید.*/}
-          {/*</Typography>*/}
-
-          <SambaUsersTable
-            users={sambaUsers}
-            isLoading={sambaUsersQuery.isLoading || sambaUsersQuery.isFetching}
-            error={sambaUsersQuery.error ?? null}
-            selectedUsers={selectedSambaUsers}
-            onToggleSelect={handleToggleSelectSambaUser}
-            onEnable={handleEnableSambaUser}
-            onEditPassword={handleOpenPasswordModal}
-            pendingEnableUsername={pendingEnableUsername}
-            isEnabling={enableSambaUser.isPending}
-            pendingPasswordUsername={pendingPasswordUsername}
-            isUpdatingPassword={updateSambaPassword.isPending}
-          />
-
-          <SelectedSambaUsersDetailsPanel
-            items={selectedSambaUserItems}
-            onRemove={handleRemoveSelectedSambaUser}
-          />
-        </Box>
-      </TabPanel>
-
       <TabPanel value={USERS_TABS.other} currentValue={activeTab}>
         <Typography sx={{ color: 'var(--color-secondary)' }}>
           بخش سایر کاربران در دست توسعه است.
@@ -512,15 +318,6 @@ const Users = () => {
         isSubmitting={createSambaUser.isPending}
         errorMessage={sambaCreateError}
         initialUsername={sambaCreateInitialUsername}
-      />
-
-      <SambaUserPasswordModal
-        open={isPasswordModalOpen}
-        username={passwordModalUsername}
-        onClose={handleClosePasswordModal}
-        onSubmit={handleSubmitPasswordChange}
-        isSubmitting={updateSambaPassword.isPending}
-        errorMessage={passwordModalError}
       />
     </Box>
   );
